Configuration

These functions are used at the beginning of your script to set up the configuration.

Main.Finch.domain

Main.Finch.functionSpace

Main.Finch.generateFor

Main.Finch.matrixFree

Main.Finch.nodeType

Main.Finch.setSteps

Main.Finch.solverType

Main.Finch.timeInterval

Main.Finch.timeStepper

Main.Finch.useLog

Main.Finch.generateFor — Method

generateFor(lang; filename=project_name, header="", params=nothing)

Specify the generation target. Lang could be one of the included target constants: (MATLAB, DENDRO) or the filename where the target is defined. The keyword argument filename refers to the name to be applied to the generated code. The header text will be placed at the top of each generated code file. If the target requires some extra parameters, those are included in params.

Main.Finch.useLog — Function

useLog(name=project_name; dir=output_dir, level=2)

Turn on logging with the given file name and optional directory. The verbosity level can be 1(basic progress info), 2(More details about each step), or 3(everything).

Main.Finch.domain — Method

domain(dims; shape=SQUARE, grid=UNIFORM_GRID)

Set the dimensionality of the domain. The shape(SQUARE, IRREGULAR) and grid type(UNIFORM_GRID, UNSTRUCTURED, TREE) can be set, but may be changed when building or importing the mesh.

Main.Finch.solverType — Function

solverType(method, backend=DEFAULT_SOLVER)

Select between CG, DG, FV, or MIXED methods. The backend refers to the tools used for solving linear systems: DEFAULTSOLVER, PETSCSOLVER, CUDA_SOLVER. Not all combinations of options are available yet.

Main.Finch.functionSpace — Method

functionSpace(;space=LEGENDRE, order=0, orderMin=0, orderMax=0)

Set the polynomial order and type of polynomials for FEM. Some of these are placeholders, so only use order at this point.

Main.Finch.nodeType — Method

nodeType(type)

For FEM, set the nodal configuration within elements. The default is LOBATTO. GAUSS and UNIFORM are available, but should be used with care.

Main.Finch.timeStepper — Method

timeStepper(type; cfl=0)

Set the type of time stepping method and optionally the CFL number. Options include EULEREXPLICIT, EULERIMPLICIT, CRANK_NICHOLSON, RK4, LSRK4, PECE. If no CFL number is provided, one will be chosen based on the mesh and stepper type.

Main.Finch.timeInterval — Method

timeInterval(T)

Set the ending time for time stepping. This is overridden if time steps are manually specified.

Main.Finch.setSteps — Method

setSteps(dt, steps)

Manually set the time steps if desired.

Main.Finch.matrixFree — Function

matrixFree(shallwe=true; maxiters=100, tol=1e-6)

Select a matrix free method for FEM with the given max iterations and tolerance. This will use a basic conjugate gradient method.