Equation Input

Functions related to the PDE expressions.

Finch.flux

Finch.printLatex

Finch.source

Finch.weakForm

Finch.weakForm — Method

weakForm(var, wf)

Write the weak form of the PDE in residual form. This should be an expression that is assumed to be equal to zero. var can be a variable or an array of variables. When using arrays, wf must also be an array of matching size. wf is a string expression or array of them. It can include numbers, coefficients, variables, parameters, indexers, and symbolic operators.

Finch.flux — Method

flux(var, fex)

Write the surface integral term of the PDE. This is the flux after transforming using the divergence theorem into a surface integral. var can be a variable or an array of variables. When using arrays, fex must also be an array of matching size. fex is a string expression or array of them. It can include numbers, coefficients, variables, parameters, indexers, and symbolic operators.

Finch.source — Method

source(var, sex)

Write the volume integral term of the PDE. var can be a variable or an array of variables. When using arrays, fex must also be an array of matching size. sex is a string expression or array of them. It can include numbers, coefficients, variables, parameters, indexers, and symbolic operators.

Finch.printLatex — Method

printLatex(var)

Print a string of Latex formatted code for the symbolic layer form of the PDE. This is somewhat limited and needs to be updated to a more useful output.