Mesh

These functions work with the mesh.

Main.Finch.exportMesh

Main.Finch.finiteVolumeOrder

Main.Finch.mesh

Main.Finch.mesh — Method

mesh(msh; elsperdim=5, bids=1, interval=[0,1], partitions=0)

Build or import a mesh. msh can be either a constant(LINEMESH, QUADMESH, HEXMESH) to build a mesh with the built-in simple mesh generator, or a filename for a mesh file. Currently GMSH files(.msh), either old or new versions, and MEDIT files(.mesh) are supported.

Main.Finch.exportMesh — Function

exportMesh(filename, format=MSH_V2)

Export the mesh to a file with the given name. The format can be MSHV2 or MSHV4 for old and new style GMSH formats.

Main.Finch.finiteVolumeOrder — Method

finiteVolumeOrder(order)

Set the order of flux reconstruction for FVM. For order > 1 this will cause the mesh to be subdivided into a parent/child mesh. Take this into account when designing the mesh.