Solution and Output

Solve and output the results.

Finch.outputValues

Finch.solve

Finch.solve — Method

solve(var)

Either solve the problem using an internal target, or generate all code files for an external target. Var is a variable or array of variables to solve for. The keyword arguments are for nonlinear equations which have very limited support, so generally they won't be used.

Finch.outputValues — Method

outputValues(vars, filename; format="vtk", ascii=false)

Output variable values to a file in a spicified format. vars can be a variable or array of variables. Possible formats are "vtk", "csv", or "raw" Set ascii to true to make ascii type vtk files instead of binary.