Other Interface Functions

These miscellaneous parts are part of the interface, but don't fit in the other categories.

Finch.VariableTransform

Finch.assemblyLoops

Finch.cachesim

Finch.cachesimSolve

Finch.callbackFunction

Finch.customOperator

Finch.customOperatorFile

Finch.exportCode

Finch.finalizeFinch

Finch.importCode

Finch.initFinch

Finch.postStepFunction

Finch.preStepFunction

Finch.transformVariable

Finch.variableTransform

Finch.initFinch — Function

initFinch(name = "unnamedProject")

This initializes the Finch state. If Finch had been previously initialized, this will reset everything. The name of the project can be set here.

Finch.customOperator — Method

customOperator(name, handle)

Define a new symbolic operator to be used in PDE expressions. The name is the symbol that will be used in expressions. The function handle points to the operator's function.

Finch.customOperatorFile — Method

customOperatorFile(filename)

Import a set of symbolic operators defined in a file. The file must contain certain elements. See an example.

Finch.VariableTransform — Type

VariableTransform

Two sets of variables and a function that transforms one into the other. This is built with the variableTransform function.

Finch.variableTransform — Method

variableTransform(var1, var2, func)

Define a function that will transform var1 into var2. var1 and var2 can be variable entities or arrays of them. The function should transform numerical values from one to the other.

Finch.transformVariable — Method

transformVariable(xform::VariableTransform)

Perform a variable transformation using a previously defined transform. This can be done, for example, in a postStepFunction or in a callback function for boundary conditions.

Finch.preStepFunction — Method

preStepFunction(fun)

Set a function to be called before each time step, or stage for multi-stage steppers.

Finch.postStepFunction — Method

postStepFunction(fun)

Set a function to be called after each time step, or stage for multi-stage steppers.

Finch.callbackFunction — Method

callbackFunction(fun; name="", args=[], body="")

Include a callback function that can be included in expressions such as the PDE or boundary conditions. A better way to do this is with the macro @callbackFunction before the function definition, which will extract the name, arguments, and body automatically.

Finch.assemblyLoops — Function

assemblyLoops(indices, parallel_type=[])

Specify the nesting order of loops for assembling the system. This makes sense for problems with indexed variables. Var is a variable of array of variables. Indices is an array of indexer objects and a string "elements". Loops will be nested in that order with outermost first. If paralleltype is specified for the loops, they will be generated with that type of parallel strategy. Possibilities are "none", "mpi", "threads". NOTE: paralleltype use is still under development and may not work for all problems.

Finch.exportCode — Method

exportCode(filename)

Export all generated code including the elemental calculation and assembly loop code for all variables if available.

Finch.importCode — Method

importCode(filename)

Import elemental calculation and assembly loop code for all variables if possible. The "begin", "end", and "No code" comment lines must match a specific format to properly match them to the variables, so do not modify those lines from the exported code.

Finch.finalizeFinch — Method

finalizeFinch()

This closes all files and does any other finalization steps. It does not deallocate any data, so further processing can be done after calling this. However, using Finch functions may cause issues because files have been closed.

Finch.cachesim — Method

cachesim(use)

Toggle the cache simulator target. Set use=true to use the cachesim target. If not using cachesim, don't use this function.

Finch.cachesimSolve — Method

cachesimSolve(var)

When using the cache simulator target, this is used instead of solve().