Reordering

Reorder the elemental loop or nodes. Most of these assume a uniform grid of squares or cubes such as those produced by the internal mesh creation utilities.

Main.Finch.elementFirstNodes

Main.Finch.hilbertElements

Main.Finch.hilbertNodes

Main.Finch.mortonElements

Main.Finch.mortonNodes

Main.Finch.randomElements

Main.Finch.randomNodes

Main.Finch.tiledElements

Main.Finch.tiledNodes

Main.Finch.mortonElements — Method

mortonElements(griddim)

Reorder the elemental loop order to a spacial Morton ordering in 2D or 3D. This currently only works for a uniform grid such as the one generated with Finch's internal utility. griddim is an array representing the elemental grid size: like [n,n] for 2D or [n,n,n] for 3D.

Main.Finch.hilbertElements — Method

hilbertElements(griddim)

Reorder the elemental loop order to a spacial Hilbert ordering in 2D or 3D. This currently only works for a uniform grid such as the one generated with Finch's internal utility. griddim is an array representing the elemental grid size: like [n,n] for 2D or [n,n,n] for 3D.

Main.Finch.tiledElements — Method

tiledElements(griddim, tiledim)

Reorder the elemental loop order to a spacial tiled ordering in 2D or 3D. This currently only works for a uniform grid such as the one generated with Finch's internal utility. griddim is an array representing the elemental grid size: like [n,n] for 2D or [n,n,n] for 3D. tiledim is the desired tile dimensions such as [4,4] for a 4x4 tile in 2D.

Main.Finch.randomElements — Function

randomElements(seed = 17)

Randomize the order of the elemental loop for testing a worst-case arrangement. The seed is for making results reproducible.

Main.Finch.mortonNodes — Method

mortonNodes(griddim)

Reorder the nodes in memory to a Morton ordering in 2D or 3D. This currently only works for a uniform grid such as the one generated with Finch's internal utility. griddim is an array representing the nodal grid size: like [n,n] for 2D or [n,n,n] for 3D.

Main.Finch.hilbertNodes — Method

hilbertNodes(griddim)

Reorder the nodes in memory to a Hilbert ordering in 2D or 3D. This currently only works for a uniform grid such as the one generated with Finch's internal utility. griddim is an array representing the nodal grid size: like [n,n] for 2D or [n,n,n] for 3D.

Main.Finch.tiledNodes — Method

tiledNodes(griddim, tiledim)

Reorder the nodes in memory to a tiled ordering in 2D or 3D. This currently only works for a uniform grid such as the one generated with Finch's internal utility. griddim is an array representing the nodal grid size: like [n,n] for 2D or [n,n,n] for 3D. tiledim is the desired tile dimensions such as [4,4] for a 4x4 tile in 2D.

Main.Finch.elementFirstNodes — Method

elementFirstNodes()

This is the default node ordering. Element first means the elements are given some order and the nodes are added elementwise according to that. An element's nodes are ordered according to the reference element.

Main.Finch.randomNodes — Function

randomNodes(seed = 17)

Randomize nodes in memory for testing a worst-case arrangement. The seed is for making results reproducible.